code divided into classes. async functions refactored, with shorter code
