+++ src/group_tools.py
@@ -1,0 +1,33 @@
+from operator import itemgetter
+
+
+class GroupTools:
+
+    def __init__(self, client, **kwargs):
+        self.client = client
+        self.group_id = kwargs.get('group_id', None)
+
+    async def list_all_users(self):
+        for participant in await self.client.get_participants(self.group_id, aggressive=True):
+            participant = {"id": participant.id,
+                           "first_name": participant.first_name,
+                           "last_name": participant.last_name,
+                           "count": 0}
+            yield participant
+
+    async def count_messages(self):
+        participants_data = [p async for p in self.list_all_users()]
+        async for message in self.client.iter_messages(self.group_id):
+            for participant in participants_data:
+                if participant['id'] == message.from_id:
+                    participant['count'] += 1
+        participants_sorted = sorted(participants_data, key=itemgetter('count'), reverse=True)
+        return participants_sorted
+
+    async def count_messages_of_specific_user(self, **kwargs):
+        # based on name or user_id
+        pass
+
+    async def get_messages_with_content(self, **kwargs):
+        # TODO
+        pass
